+++ package.json
@@ -36,6 +36,7 @@
 		"eslint": "^3.15.0",
 		"eslint-plugin-babel": "^4.0.1",
 		"eslint-plugin-import": "^2.2.0",
+		"fmt-obj": "^1.3.0",
 		"nodemon": "^1.11.0"
 	},
 	"ava": {
