+++ server/resolvers/index.js
@@ -1,5 +1,5 @@
 import { ChatroomResolve, MessageResolve } from './associations';
-import { chatrooms, messages, users } from './query';
+import { chatrooms, chatroom, users, user, messages } from './query';
 
 import { addMessage } from './mutation';
 import { messageAdded } from './subscription';
@@ -7,7 +7,9 @@ import { messageAdded } from './subscription';
 const resolvers = {
 	Query: {
 		chatrooms,
+		chatroom,
 		users,
+		user,
 		messages
 	},
 	Chatroom: ChatroomResolve,
