+++ server/resolvers/mutation/index.js
@@ -1,5 +1,5 @@
 import { Message } from '../../models/index';
-import { pubsub } from '../../subscriptions';
+// import { pubsub } from '../../subscriptions';
 
 const addMessage = (obj, args, context) => {
 	return Message.create({
@@ -7,7 +7,7 @@ const addMessage = (obj, args, context) => {
 		userId: args.userId,
 		chatroomId: args.chatroomId
 	}).then(message => {
-		pubsub.publish('newMessagesChannel', message.dataValues);
+		// pubsub.publish('newMessagesChannel', message.dataValues);
 		return message.dataValues
 	});
 }
